+++ MikuMikuWorld/ScorePreview.h
@@ -52,7 +52,7 @@ namespace MikuMikuWorld
 		
 		void drawNoteBase(Renderer* renderer, const Note& note, float left, float right, float y, float zScalar = 1);
 		void drawTraceDiamond(Renderer* renderer, const Note& note, float left, float right, float y);
-		void drawFlickArrow(Renderer* renderer, const Note& note, float y, float cur_time);
+		void drawFlickArrow(Renderer* renderer, const Note& note, float y, double cur_time);
 		void drawParticle(Renderer* renderer, const std::array<DirectX::XMFLOAT4, 4>& layout, const Engine::DrawingParticle& particle,
 			float progress, const Texture& texture, const Sprite& sprite, int zIndex, Color tint);
